requirements.txt
# ═══════════════════════════════════════════════════════════════
#  Generative Fashion Designer — Dependencies
# ═══════════════════════════════════════════════════════════════

# Core DL
torch>=2.2
torchvision>=0.17
numpy>=1.24

# Config
pyyaml>=6.0

# Visualization
matplotlib>=3.7
seaborn>=0.12

# ML utilities
scikit-learn>=1.3
tqdm>=4.65

# TensorBoard
tensorboard>=2.14
lightning>=2.3

# Image processing
Pillow>=10.0
